Additional CR changes.
Note - FlightSqlExample is not functional and needs to be updated.
