+++ java/flight/flight-sql/src/main/java/org/apache/arrow/flight/sql/FlightSqlUtils.java
@@ -17,14 +17,9 @@
 
 package org.apache.arrow.flight.sql;
 
-import java.sql.Types;
 import java.util.List;
 
 import org.apache.arrow.flight.ActionType;
-import org.apache.arrow.vector.types.DateUnit;
-import org.apache.arrow.vector.types.FloatingPointPrecision;
-import org.apache.arrow.vector.types.TimeUnit;
-import org.apache.arrow.vector.types.pojo.ArrowType;
 
 import com.google.common.collect.ImmutableList;
 import com.google.protobuf.Any;
@@ -35,14 +30,6 @@ import com.google.protobuf.Message;
  * Utilities to work with Flight SQL semantics.
  */
 public final class FlightSqlUtils {
-
-  private static final int BIT_WIDTH8 = 8;
-  private static final int BIT_WIDTH_16 = 16;
-  private static final int BIT_WIDTH_32 = 32;
-  private static final int BIT_WIDTH_64 = 64;
-  private static final boolean IS_SIGNED_FALSE = false;
-  private static final boolean IS_SIGNED_TRUE = true;
-
   public static final ActionType FLIGHT_SQL_CREATEPREPAREDSTATEMENT = new ActionType("CreatePreparedStatement",
           "Creates a reusable prepared statement resource on the server. \n" +
                   "Request Message: ActionCreatePreparedStatementRequest\n" +
@@ -57,76 +44,6 @@ public final class FlightSqlUtils {
           FLIGHT_SQL_CREATEPREPAREDSTATEMENT,
           FLIGHT_SQL_CLOSEPREPAREDSTATEMENT
   );
-
-  /**
-   * Converts {@link java.sql.Types} values returned from JDBC Apis to Arrow types.
-   *
-   * @param jdbcDataType {@link java.sql.Types} value.
-   * @param precision    Precision of the type.
-   * @param scale        Scale of the type.
-   * @return The Arrow equivalent type.
-   */
-  public static ArrowType getArrowTypeFromJDBCType(int jdbcDataType, int precision, int scale) {
-
-    switch (jdbcDataType) {
-      case Types.BIT:
-      case Types.BOOLEAN:
-        return ArrowType.Bool.INSTANCE;
-      case Types.TINYINT:
-        return new ArrowType.Int(BIT_WIDTH8, IS_SIGNED_TRUE);
-      case Types.SMALLINT:
-        return new ArrowType.Int(BIT_WIDTH_16, IS_SIGNED_TRUE);
-      case Types.INTEGER:
-        return new ArrowType.Int(BIT_WIDTH_32, IS_SIGNED_TRUE);
-      case Types.BIGINT:
-        return new ArrowType.Int(BIT_WIDTH_64, IS_SIGNED_TRUE);
-      case Types.FLOAT:
-      case Types.REAL:
-        return new ArrowType.FloatingPoint(FloatingPointPrecision.SINGLE);
-      case Types.DOUBLE:
-        return new ArrowType.FloatingPoint(FloatingPointPrecision.DOUBLE);
-      case Types.NUMERIC:
-      case Types.DECIMAL:
-        return new ArrowType.Decimal(precision, scale);
-      case Types.DATE:
-        return new ArrowType.Date(DateUnit.DAY);
-      case Types.TIME:
-        return new ArrowType.Time(TimeUnit.MILLISECOND, BIT_WIDTH_32);
-      case Types.TIMESTAMP:
-        return new ArrowType.Timestamp(TimeUnit.MILLISECOND, null);
-      case Types.BINARY:
-      case Types.VARBINARY:
-      case Types.LONGVARBINARY:
-        return ArrowType.Binary.INSTANCE;
-      case Types.NULL:
-        return ArrowType.Null.INSTANCE;
-
-      case Types.CHAR:
-      case Types.VARCHAR:
-      case Types.LONGVARCHAR:
-      case Types.CLOB:
-      case Types.NCHAR:
-      case Types.NVARCHAR:
-      case Types.LONGNVARCHAR:
-      case Types.NCLOB:
-
-      case Types.OTHER:
-      case Types.JAVA_OBJECT:
-      case Types.DISTINCT:
-      case Types.STRUCT:
-      case Types.ARRAY:
-      case Types.BLOB:
-      case Types.REF:
-      case Types.DATALINK:
-      case Types.ROWID:
-      case Types.SQLXML:
-      case Types.REF_CURSOR:
-      case Types.TIME_WITH_TIMEZONE:
-      case Types.TIMESTAMP_WITH_TIMEZONE:
-      default:
-        return ArrowType.Utf8.INSTANCE;
-    }
-  }
 
   /**
    * Helper to parse {@link com.google.protobuf.Any} objects to the specific protobuf object.
