+++ java/flight/flight-sql/src/test/java/org/apache/arrow/flight/TestFlightSql.java
@@ -17,42 +17,29 @@
 
 package org.apache.arrow.flight;
 
-import static org.apache.arrow.flight.sql.FlightSqlClientUtils.getPreparedStatement;
-import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertFalse;
-import static org.junit.Assert.assertTrue;
-
-import java.nio.charset.StandardCharsets;
+import java.nio.ByteBuffer;
 import java.util.ArrayList;
 import java.util.Arrays;
-import java.util.Iterator;
 import java.util.List;
 
-import org.apache.arrow.flight.sql.FlightSqlClientUtils;
+import org.apache.arrow.flight.sql.FlightSqlClient;
 import org.apache.arrow.flight.sql.FlightSqlExample;
-import org.apache.arrow.flight.sql.impl.FlightSql.ActionClosePreparedStatementRequest;
-import org.apache.arrow.flight.sql.impl.FlightSql.ActionCreatePreparedStatementRequest;
-import org.apache.arrow.flight.sql.impl.FlightSql.ActionCreatePreparedStatementResult;
-import org.apache.arrow.flight.sql.impl.FlightSql.CommandPreparedStatementQuery;
 import org.apache.arrow.memory.BufferAllocator;
 import org.apache.arrow.memory.RootAllocator;
-import org.apache.arrow.memory.util.ArrowBufPointer;
 import org.apache.arrow.util.AutoCloseables;
 import org.apache.arrow.vector.FieldVector;
 import org.apache.arrow.vector.IntVector;
+import org.apache.arrow.vector.VarBinaryVector;
 import org.apache.arrow.vector.VarCharVector;
 import org.apache.arrow.vector.VectorSchemaRoot;
-import org.apache.arrow.vector.types.pojo.ArrowType;
+import org.apache.arrow.vector.types.Types;
 import org.apache.arrow.vector.types.pojo.Field;
 import org.apache.arrow.vector.types.pojo.FieldType;
 import org.apache.arrow.vector.types.pojo.Schema;
-import org.apache.arrow.vector.util.ElementAddressableVectorIterator;
 import org.junit.AfterClass;
 import org.junit.BeforeClass;
 import org.junit.Test;
-
-import com.google.protobuf.Any;
-import com.google.protobuf.ByteString;
+import org.junit.jupiter.api.Assertions;
 
 /**
  * Test direct usage of Flight SQL workflows.
@@ -62,14 +49,11 @@ public class TestFlightSql {
   private static FlightServer server;
 
   private static FlightClient client;
+  private static FlightSqlClient sqlClient;
 
   protected static final Schema SCHEMA_INT_TABLE = new Schema(Arrays.asList(
-          new Field("KEYNAME", new
-                  FieldType(true, ArrowType.Utf8.INSTANCE, null),
-                  null),
-          new Field("VALUE",
-                  new FieldType(true, new ArrowType.Int(32, true), null),
-                  null)));
+          new Field("KEYNAME", FieldType.nullable(Types.MinorType.VARCHAR.getType()), null),
+          new Field("VALUE", FieldType.nullable(Types.MinorType.INT.getType()), null)));
 
   @BeforeClass
   public static void setUp() throws Exception {
@@ -81,6 +65,7 @@ public class TestFlightSql {
 
     final Location clientLocation = Location.forGrpcInsecure(FlightTestUtil.LOCALHOST, server.getPort());
     client = FlightClient.builder(allocator, clientLocation).build();
+    sqlClient = new FlightSqlClient(client);
   }
 
   @AfterClass
@@ -88,173 +73,73 @@ public class TestFlightSql {
     AutoCloseables.close(client, server, allocator);
   }
 
-  /*
   @Test
   public void testGetTables() throws Exception {
-    // Arrange
-    final ActionGetTablesResult expected = ActionGetTablesResult.newBuilder()
-            .setSchema("APP")
-            .setTable("INTTABLE")
-            .setTableType("TABLE")
-            .setArrowMetadata(ByteString.copyFrom(SCHEMA_INT_TABLE.toByteArray()))
-            .build();
-
-    // Act
-    final Iterator<Result> results = client.doAction(new Action("GetTables",
-            Any.pack(ActionGetTablesRequest
-                    .newBuilder()
-                    .addTableTypes("TABLE")
-                    .setIncludeSchema(true)
-                    .build())
-                    .toByteArray()));
-
-    // Assert
-    while (results.hasNext()) {
-      ActionGetTablesResult actual = Any.parseFrom(results.next().getBody()).unpack(ActionGetTablesResult.class);
-      assertEquals(expected, actual);
+    final FlightInfo info = sqlClient.getTables(null, null, null, null, false);
+    try (final FlightStream stream = sqlClient.getStream(info.getEndpoints().get(0).getTicket())) {
+      final List<List<String>> results = getResults(stream);
+      Assertions.assertEquals(1, results.size());
+      Assertions.assertEquals(
+          Arrays.asList(null, "APP", "INTTABLE", "TABLE", SCHEMA_INT_TABLE.toJson()), results.get(0));
     }
   }
-
-  @Test
-  public void testGetTablesWithFlightSqlClientUtils() throws Exception {
-    // Arrange
-    final ActionGetTablesResult expected = ActionGetTablesResult.newBuilder()
-            .setSchema("APP")
-            .setTable("INTTABLE")
-            .setTableType("TABLE")
-            .setArrowMetadata(ByteString.copyFrom(SCHEMA_INT_TABLE.toByteArray()))
-            .build();
-
-    // Act
-    final List<ActionGetTablesResult> results = getTables(client, null, null, null,
-            Collections.singletonList("TABLE"), true);
-
-    // Assert
-    assertEquals(1, results.size());
-    assertEquals(expected, results.get(0));
-  }
-  */
 
   @Test
   public void testSimplePrepStmt() throws Exception {
-    final Iterator<Result> preparedStatementResults = client.doAction(new Action("GetPreparedStatement",
-            Any.pack(ActionCreatePreparedStatementRequest
-                    .newBuilder()
-                    .setQuery("Select * from intTable")
-                    .build())
-                    .toByteArray()));
-
-    assertTrue(preparedStatementResults.hasNext());
-    final ActionCreatePreparedStatementResult preparedStatementResult =
-            Any.parseFrom(preparedStatementResults.next().getBody()).unpack(ActionCreatePreparedStatementResult.class);
-    assertFalse(preparedStatementResults.hasNext());
-
-    final Schema actualSchema = Schema.deserialize(preparedStatementResult.getDatasetSchema().asReadOnlyByteBuffer());
-    assertEquals(SCHEMA_INT_TABLE, actualSchema);
-
-    final FlightDescriptor descriptor = FlightDescriptor
-            .command(Any.pack(CommandPreparedStatementQuery.newBuilder()
-                    .setClientExecutionHandle(ByteString.copyFrom(new byte[]{1, 2, 3, 4}))
-                    .setPreparedStatementHandle(preparedStatementResult.getPreparedStatementHandle())
-                    .build())
-                    .toByteArray());
-
-    final FlightInfo info = client.getInfo(descriptor);
-    assertEquals(SCHEMA_INT_TABLE, info.getSchema());
-
-    final FlightStream stream = client.getStream(info.getEndpoints().get(0).getTicket());
-    assertEquals(SCHEMA_INT_TABLE, stream.getSchema());
-
-    List<String> actualStringResults = new ArrayList<>();
-    List<Integer> actualIntResults = new ArrayList<>();
-    while (stream.next()) {
-      final VectorSchemaRoot root = stream.getRoot();
-      final long rowCount = root.getRowCount();
-
-      for (Field field : root.getSchema().getFields()) {
-        final FieldVector fieldVector = root.getVector(field.getName());
-
-        if (fieldVector instanceof VarCharVector) {
-
-          final ElementAddressableVectorIterator<VarCharVector> it =
-                  new ElementAddressableVectorIterator<>((VarCharVector) fieldVector);
-
-          for (int rowIndex = 0; rowIndex < rowCount; rowIndex++) {
-            final ArrowBufPointer pt = it.next();
-            final byte[] bytes = new byte[(int) pt.getLength()];
-            pt.getBuf().getBytes(pt.getOffset(), bytes);
-
-            actualStringResults.add(new String(bytes, StandardCharsets.UTF_8));
-          }
-        } else if (fieldVector instanceof IntVector) {
-          for (int rowIndex = 0; rowIndex < rowCount; rowIndex++) {
-            actualIntResults.add(((IntVector) fieldVector).get(rowIndex));
-          }
-        }
-      }
-    }
-    stream.getRoot().clear();
-
-    assertEquals(Arrays.asList("one", "zero", "negative one"), actualStringResults);
-    assertEquals(Arrays.asList(1, 0, -1), actualIntResults);
-
-    final Iterator<Result> closePreparedStatementResults = client.doAction(new Action("ClosePreparedStatement",
-            Any.pack(ActionClosePreparedStatementRequest
-                    .newBuilder()
-                    .setPreparedStatementHandleBytes(preparedStatementResult.getPreparedStatementHandle())
-                    .build())
-                    .toByteArray()));
-    assertFalse(closePreparedStatementResults.hasNext());
-  }
-
-  @Test
-  public void testSimplePrepStmtWithFlightSqlClientUtils() throws Exception {
-    final FlightSqlClientUtils.FlightSqlPreparedStatement preparedStatement =
-            getPreparedStatement(client, "Select * from intTable");
+    final FlightSqlClient.PreparedStatement preparedStatement = sqlClient.prepare("Select * from intTable");
 
     final Schema actualSchema = preparedStatement.getResultSetSchema();
-    assertEquals(SCHEMA_INT_TABLE, actualSchema);
+    Assertions.assertEquals(SCHEMA_INT_TABLE, actualSchema);
 
-    final FlightInfo info = preparedStatement.executeQuery();
-    assertEquals(SCHEMA_INT_TABLE, info.getSchema());
+    final FlightInfo info = preparedStatement.execute();
+    Assertions.assertEquals(SCHEMA_INT_TABLE, info.getSchema());
 
-    final FlightStream stream = client.getStream(info.getEndpoints().get(0).getTicket());
-    assertEquals(SCHEMA_INT_TABLE, stream.getSchema());
+    try (final FlightStream stream = sqlClient.getStream(info.getEndpoints().get(0).getTicket())) {
+      Assertions.assertEquals(SCHEMA_INT_TABLE, stream.getSchema());
 
-    List<String> actualStringResults = new ArrayList<>();
-    List<Integer> actualIntResults = new ArrayList<>();
+      final List<List<String>> results = getResults(stream);
+      Assertions.assertEquals(3, results.size());
+      Assertions.assertEquals(Arrays.asList("one", "1"), results.get(0));
+      Assertions.assertEquals(Arrays.asList("zero", "0"), results.get(1));
+      Assertions.assertEquals(Arrays.asList("negative one", "-1"), results.get(2));
+    }
+
+    AutoCloseables.close(preparedStatement);
+    Assertions.assertTrue(preparedStatement.isClosed());
+  }
+
+  List<List<String>> getResults(FlightStream stream) {
+    final List<List<String>> results = new ArrayList<>();
     while (stream.next()) {
       final VectorSchemaRoot root = stream.getRoot();
       final long rowCount = root.getRowCount();
+      for (int i = 0; i < rowCount; ++i) {
+        results.add(new ArrayList<>());
+      }
 
       for (Field field : root.getSchema().getFields()) {
         final FieldVector fieldVector = root.getVector(field.getName());
 
         if (fieldVector instanceof VarCharVector) {
-
-          final ElementAddressableVectorIterator<VarCharVector> it =
-                  new ElementAddressableVectorIterator<>((VarCharVector) fieldVector);
-
+          final VarCharVector varcharVector = (VarCharVector) fieldVector;
           for (int rowIndex = 0; rowIndex < rowCount; rowIndex++) {
-            final ArrowBufPointer pt = it.next();
-            final byte[] bytes = new byte[(int) pt.getLength()];
-            pt.getBuf().getBytes(pt.getOffset(), bytes);
-
-            actualStringResults.add(new String(bytes, StandardCharsets.UTF_8));
+            results.get(rowIndex).add(varcharVector.getObject(rowIndex).toString());
           }
         } else if (fieldVector instanceof IntVector) {
           for (int rowIndex = 0; rowIndex < rowCount; rowIndex++) {
-            actualIntResults.add(((IntVector) fieldVector).get(rowIndex));
+            results.get(rowIndex).add(String.valueOf(((IntVector) fieldVector).get(rowIndex)));
           }
+        } else if (fieldVector instanceof VarBinaryVector) {
+          final VarBinaryVector varbinaryVector = (VarBinaryVector) fieldVector;
+          for (int rowIndex = 0; rowIndex < rowCount; rowIndex++) {
+            results.get(rowIndex).add(Schema.deserialize(ByteBuffer.wrap(varbinaryVector.get(rowIndex))).toJson());
+          }
+        } else {
+          throw new UnsupportedOperationException("Not yet implemented");
         }
       }
     }
     stream.getRoot().clear();
-
-    assertEquals(Arrays.asList("one", "zero", "negative one"), actualStringResults);
-    assertEquals(Arrays.asList(1, 0, -1), actualIntResults);
-
-    AutoCloseables.close(preparedStatement);
-    assertTrue(preparedStatement.isClosed());
+    return results;
   }
 }
